[nacltoons] Add linux host build.
Also, add some easing actions to level selection menu as an example
of how to so node animations.

Review URL: https://codereview.chromium.org/12700018

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1534 050acbb0-2703-11df-ab0a-9f3f633ae91d

+++ nacltoons/bindings/LuaBox2D.cpp
@@ -7033,7 +7033,7 @@ static int tolua_LuaBox2D_b2Body_SetUserData01(lua_State* tolua_S)
   if (!self) tolua_error(tolua_S,"invalid 'self' in function 'SetUserData'", NULL);
 #endif
   {
-   self->SetUserData((void*)data);
+   self->SetUserData((void*)(long)data);
   }
  }
  return 0;
